Making base.sql.registry its own module
