+++ ibis/backends/base_sql/compiler.py
@@ -1,6 +1,9 @@
 import ibis.backends.base_sqlalchemy.compiler as comp
 import ibis.expr.operations as ops
-from ibis.backends.base.sql import operation_registry, quote_identifier
+from ibis.backends.base.sql.registry import (
+    operation_registry,
+    quote_identifier,
+)
 
 
 def build_ast(expr, context):
