+++ ibis/backends/spark/registry.py
@@ -7,7 +7,7 @@ import ibis.expr.datatypes as dt
 import ibis.expr.operations as ops
 import ibis.expr.types as ir
 import ibis.util as util
-from ibis.backends.base.sql import (
+from ibis.backends.base.sql.registry import (
     fixed_arity,
     literal_formatters,
     operation_registry,
